Create Vridhi_IPL_T20_Cricket_Analysis_Capstone_Project.ipynb
So what can we expect from the multiple IPL datasets?¶
The datasets are:
1.matches.csv
2.Players.xlsx
3.deliveries.csv
4.teams.csv
5.most_runs_average_strikerate.csv
6.teamwise_home_and_away.csv
In order to understand our data, we can look at each variable and try to understand their meaning and relevance to this problem.
The datsets consists of multiple observations & columns. 
Let's go ahead & get familiar with the columns.

Cleaning the dataset

From matches dataset we have:-ID: 
Basically the IPL match ID.
Season: 
The IPL season, which in our dataset if from season 2008 - season 2019
City:The cities where the IPL matches were played.
Date: The date on which the matches were played or held.
Team1 & Team2: The teams participated in the IPL matches since 2008.Toss_winner: The team that won the toss.
Toss_decision: The decision taken by the toss winning team either to 'Bat' or 'Field' first.
Result: The result column consists of (normal, tie, no result) of the match.
Dl_applied: The column consists the numerical values (0 or 1) which indicates whether the Duckworth-Lewis method was applied or not.
Winner: Winner of the respective match.
Win_by_runs: The runs by which the team batting first won.
Win_by_wickets: The wickets by which the team batting second won.
Player_of_the_match: The player who performed exceptionally good in the match.
Venue: The venue where the match was hosted.
Umpire1 & Umpire2: Two on-field umpires during the match.Umpire3: Basically the third Umpire.